requirements.txt
streamlit
python-dotenv
huggingface_hub
requests
PIL
requests
openai